User routes refactor and get user by token
