+++ src/api/routes/index.js
@@ -1,9 +1,8 @@
 import express from 'express'
-import register from './register'
 import boards from './boards'
 import lists from './lists'
 import cards from './cards'
-import members from './members'
+import users from './users'
 import comments from './comments'
 
 const router = express.Router()
@@ -11,7 +10,7 @@ const router = express.Router()
 router.use('/boards', boards)
 router.use('/lists', lists)
 router.use('/cards', cards)
-router.use('/members', members)
+router.use('/users', users)
 router.use('/comments', comments)
 
 router.get('/', (req, res) => {
